Update index.html
Was able to get the image arrow on the screen by using window.onload in the View constructor. This is no good though. What happens when other
images need to be loaded? It should be able to occur when calling a prototype function. Also, the image appears in the top left corner. I think there is
a correlation somewhere.